Moved MultiErlangC to queuing subpackage
