--- pyworkforce/multi_search/__init__.py
@@ -1,3 +1,0 @@
-from pyworkforce.multi_search.erlang import MultiErlangC
-
-__all__ = ['MultiErlangC']
