+++ pyworkforce/queuing/__init__.py
@@ -1,3 +1,3 @@
-from pyworkforce.queuing.erlang import ErlangC
+from pyworkforce.queuing.erlang import ErlangC, MultiErlangC
 
-__all__ = ["ErlangC"]
+__all__ = ["ErlangC", "MultiErlangC"]
